+++ src/components/layout/Layout.js
@@ -3,10 +3,12 @@
 import React from 'react';
 import styled, { ThemeProvider } from 'styled-components';
 import { animated, useSpring } from 'react-spring';
+import Navicon from '../actions/Navicon';
 import Disclaimer from '../layout/Disclaimer';
-import Feature from '../layout/Feature';
 import Nav from '../layout/Nav';
 import SocialNav from '../layout/SocialNav';
+import Headline from '../typography/Headline';
+import patternImage from '../../images/details-pattern.png';
 import GlobalStyle from '../../utils/GlobalStyle';
 import theme from '../../utils/theme';
 import media from '../../utils/media';
@@ -15,52 +17,74 @@ const MainWrapper = styled.div`
   overflow-x: hidden;
 
   ${media.medium} {
+    display: flex;
+    height: 100%;
     overflow-x: initial;
+    width: 100%;
   }
 `;
 
 const Wrapper = styled(animated.div)`
+  background-color: ${(props) => props.theme.white};
   min-height: 100vh;
   overflow-x: hidden;
+  width: 100%;
   z-index: 1;
-`;
-
-const Content = styled.div`
-  background-color: ${props => props.theme.white};
-  border-top-left-radius: 50px;
-  border-top-right-radius: 50px;
-  margin-top: 375px;
-  padding-top: 53px;
-  position: relative;
-  z-index: 2;
 
   ${media.medium} {
-    margin-top: 70px;
-    margin-left: 50%;
-    width: 683px;
+    width: 67%;
   }
 `;
 
-const NavWrapper = styled.div`
-  padding-left: ${props => props.theme.featureSidePad};
-  padding-right: ${props => props.theme.featureSidePad};
+const Content = styled.div`
+  padding-left: 31px;
+  padding-right: 31px;
+  padding-top: 85px;
+  position: relative;
+
+  ${media.medium} {
+    background-color: ${(props) => props.theme.white};
+    min-height: 100%;
+    padding-top: 44px;
+
+    &::before {
+      background: url(${patternImage}) repeat-x center center transparent;
+      background-size: contain;
+      display: block;
+      content: '';
+      height: 18px;
+      left: 0;
+      position: absolute;
+      right: 0;
+      top: 0;
+      width: 100%;
+    }
+  }
+`;
+
+const PageHeadline = styled(Headline)`
+  color: ${(props) => props.theme.darkText};
+  padding-bottom: 32px;
+
+  ${media.large} {
+    font-size: 64px;
+  }
 `;
 
 type Props = {
-  view: string,
-  children: Array<Object>
+  headline: string,
+  children: Array<Object>,
 };
 
 function Layout(props: Props): React$Node {
   const [showSocial, setShowSocial] = React.useState(() => false);
-  const { bRad, position, xPos } = useSpring({
-    bRad: showSocial ? '40px' : '0px',
+  const { position, xPos } = useSpring({
     position: showSocial ? 'fixed' : 'relative',
-    xPos: showSocial ? 60 : 0
+    xPos: showSocial ? 60 : 0,
   });
 
   function handleToggleSocial() {
-    return setShowSocial(prevState => !prevState);
+    return setShowSocial((prevState) => !prevState);
   }
 
   return (
@@ -68,38 +92,30 @@ function Layout(props: Props): React$Node {
       <GlobalStyle />
 
       <MainWrapper>
+        <SocialNav />
+
         <Wrapper
           style={{
-            borderTopLeftRadius: bRad,
             position,
-            transform: xPos.interpolate(x => `translate3d(${x}%, 0, 0)`)
+            transform: xPos.interpolate((x) => `translate3d(${x}%, 0, 0)`),
           }}
         >
-          <Feature
-            view={props.view}
-            showSocial={showSocial}
-            onClick={handleToggleSocial}
-          />
+          <Navicon kind="dark" show={showSocial} onClick={handleToggleSocial} />
 
           <Content>
-            <NavWrapper>
-              <Nav />
-            </NavWrapper>
-
+            <PageHeadline>{props.headline}</PageHeadline>
+            <Nav />
             {props.children}
-
             <Disclaimer />
           </Content>
         </Wrapper>
-
-        <SocialNav />
       </MainWrapper>
     </ThemeProvider>
   );
 }
 
 Layout.defaultProps = {
-  view: 'main'
+  headline: '',
 };
 
 export default React.memo(Layout);
